+++ src/components/exhibition/MainCard.tsx
